spareval: move out expression evaluator and expose errors when building the expression evaluator
Possible errors:
- error converting terms from the storage
- unknown custom function
- arity error for custom function
